Split models of ophasebase into individual fields
